#11 | Restructure content files to separate medium and substack code
