+++ src/content/content_medium.js
@@ -1,0 +1,4 @@
+import './content_medium.scss';
+import { start } from './services';
+
+start();
