+++ src/content/content_substack.js
@@ -1,4 +1,4 @@
-import './content.scss';
+import './content_substack.scss';
 import { start } from './services';
 
 start();
