+++ src/content/services/index.js
@@ -4,11 +4,12 @@ import AppService from './AppService';
 import MediumService from './MediumService';
 import SubstackService from './SubstackService';
 
+const contentIsSubstack = /(^|\.)substack.com$/i.test(window.location.hostname);
+
 const { services, start } = createServices([
   ['rpc', ContentRPCService],
   ['app', AppService],
-  ['medium', MediumService],
-  ['substack', SubstackService],
+  ['integration', contentIsSubstack ? SubstackService : MediumService],
 ]);
 
 export default services;
