+++ src/manifest.js
@@ -37,13 +37,12 @@ module.exports = (browser, version) => ({
     {
       "matches": [
         "https://medium.com/**",
-        "https://*.substack.com/**"
       ],
       "js": [
-        "content/content.js"
+        "content/content_medium.js"
       ],
       "css": [
-        "content/content.css"
+        "content/content_medium.css"
       ],
       "run_at": "document_idle"
     },
@@ -52,11 +51,23 @@ module.exports = (browser, version) => ({
         "https://medium.com/media/*"
       ],
       "js": [
-        "content/content_iframe.js"
+        "content/content_medium_iframe.js"
       ],
       "all_frames": true,
       "run_at": "document_idle"
-    }
+    },
+    {
+      "matches": [
+        "https://*.substack.com/**"
+      ],
+      "js": [
+        "content/content_substack.js"
+      ],
+      "css": [
+        "content/content_substack.css"
+      ],
+      "run_at": "document_idle"
+    },
   ],
   "web_accessible_resources": ["app/*.*"],
   "permissions": [
